requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-modeltranslation==0.18.9
django-storages==1.13.2
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.14
